Let destroy cluster finish in infra retries (#36361)
Because this step always runs the installer in the background (`&`), we
need to wait for it to finish, otherwise the terraform directly
immediately gets deleted.

+++ ci-operator/step-registry/ipi/install/install/ipi-install-install-commands.sh
@@ -413,6 +413,13 @@ do
     write_install_status
     cp "${dir}"/log-bundle-*.tar.gz "${ARTIFACT_DIR}/" 2>/dev/null
     openshift-install --dir="${dir}" destroy cluster 2>&1 | grep --line-buffered -v 'password\|X-Auth-Token\|UserData:' &
+    wait "$!"
+    ret="$?"
+    if test "${ret}" -ne 0 ; then
+      echo "Failed to destroy cluster, aborting retries."
+      ret=4
+      break
+    fi
     rm -rf "$dir"
     cp -rfpv "$backup" "$dir"
   else
